CI: Run CI on pull requests and manually
Added concurrency groups to cancel jobs when a new commit is pushed, but only for PRs. On main, every commit will have all of its job complete. Prevented running the draft release step if not the master branch, as it might be problematic, especially from PRs. PRs that need to test this should be testing it on their fork, outside of PRs